pyproject.toml
[tool.poetry]
name = "dc-nft-test"
version = "0.1.0"
description = ""
authors = ["Jeevan Jamakayala <jeevan.jamakayala@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.10"
eth-brownie = "^1.18.1"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
notebook = "^6.4.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
